+++ src/ng2-table/ng2-table.component.js
@@ -10,11 +10,13 @@ var __metadata = (this && this.__metadata) || function (k, v) {
 };
 var core_1 = require("@angular/core");
 var sorter_service_1 = require("./sorter.service");
+var style_and_class_service_1 = require("./style-and-class.service");
 var Ng2TableComponent = (function () {
     function Ng2TableComponent() {
         this.rowClicked = new core_1.EventEmitter();
         this.cellItemClicked = new core_1.EventEmitter();
         this.tableConfigUpdated = new core_1.EventEmitter();
+        this.getNgThing = style_and_class_service_1.getNgThing;
     }
     // Deep copy the parts of the tableConfig that will be modified when
     // sorting and searching the columns
@@ -34,6 +36,20 @@ var Ng2TableComponent = (function () {
             columnDefs: columnDefsCopy
         });
         return tableConfigCopy;
+    };
+    Ng2TableComponent.prototype.ngAfterViewChecked = function () {
+        // update header width based on view changes
+        this.setTheadClass(this.tbody, this.thead);
+    };
+    Ng2TableComponent.prototype.setTheadClass = function (tbody, thead) {
+        if (tbody && thead) {
+            if (tbody.nativeElement.scrollHeight > tbody.nativeElement.clientHeight) {
+                thead.nativeElement.className = 'with-scrollbar';
+            }
+            else {
+                thead.nativeElement.className = 'without-scrollbar';
+            }
+        }
     };
     Ng2TableComponent.prototype.ngOnChanges = function (changes) {
         // if there's an incoming tableConfig, replace existing tableConfigCopy
@@ -171,58 +187,6 @@ var Ng2TableComponent = (function () {
         this.sortColsAdvanced(this.tableConfigCopy.columnDefs);
         this.tableConfigUpdated.emit(this.copyTableConfig(this.tableConfigCopy));
     };
-    Ng2TableComponent.prototype.getNgThing = function (thingType, classOrStyle, tableConfig, rowData, rowIndex, activeRow, col) {
-        switch (thingType + '-' + classOrStyle) {
-            case 'table-class':
-                return tableConfig['tableNgClass'] || 'table table-striped';
-            case 'table-style':
-                return tableConfig['tableNgStyle'] || '';
-            case 'row-class':
-                if (tableConfig['rowNgClassPredicate']) {
-                    return tableConfig['rowNgClassPredicate'](rowData, rowIndex, activeRow);
-                }
-                return tableConfig['rowNgClass'] || '';
-            case 'row-style':
-                if (tableConfig['rowNgStylePredicate']) {
-                    return tableConfig['rowNgStylePredicate'](rowData, rowIndex, activeRow);
-                }
-                return tableConfig['rowNgStyle'] || '';
-            case 'cell-class':
-                if (col['cellNgClassPredicate']) {
-                    return col['cellNgClassPredicate'](rowData, rowIndex, activeRow);
-                }
-                return col['cellNgClass'] || '';
-            case 'cell-style':
-                if (col['cellNgStylePredicate']) {
-                    return col['cellNgStylePredicate'](rowData, rowIndex, activeRow);
-                }
-                return col['cellNgStyle'] || '';
-            case 'cellItemButton-class':
-                if (col['cellItem']['cellItemNgClassPredicate']) {
-                    return col['cellItem']['cellItemNgClassPredicate'](rowData, rowIndex, activeRow);
-                }
-                return col['cellItem']['cellItemNgClass'] || '';
-            case 'cellItemButton-style':
-                if (col['cellItem']['cellItemNgStylePredicate']) {
-                    return col['cellItem']['cellItemNgStylePredicate'](rowData, rowIndex, activeRow);
-                }
-                return col['cellItem']['cellItemNgStyle'] || '';
-            case 'cellItemDiv-class':
-                if (col['cellItem']['cellItemNgClassPredicate']) {
-                    return col['cellItem']['cellItemNgClassPredicate'](rowData, rowIndex, activeRow);
-                }
-                return col['cellItem']['cellItemNgClass'] || '';
-            case 'cellItemDiv-style':
-                if (col['cellItem']['cellItemNgStylePredicate']) {
-                    return col['cellItem']['cellItemNgStylePredicate'](rowData, rowIndex, activeRow);
-                }
-                return col['cellItem']['cellItemNgStyle'] || '';
-            case 'header-class':
-                return col['headerNgClass'] || '';
-            case 'header-style':
-                return col['headerNgStyle'] || '';
-        }
-    };
     return Ng2TableComponent;
 }());
 __decorate([
@@ -245,11 +209,19 @@ __decorate([
     core_1.Output(),
     __metadata("design:type", Object)
 ], Ng2TableComponent.prototype, "tableConfigUpdated", void 0);
+__decorate([
+    core_1.ViewChild('tbody'),
+    __metadata("design:type", Object)
+], Ng2TableComponent.prototype, "tbody", void 0);
+__decorate([
+    core_1.ViewChild('thead'),
+    __metadata("design:type", Object)
+], Ng2TableComponent.prototype, "thead", void 0);
 Ng2TableComponent = __decorate([
     core_1.Component({
         selector: 'ng2-table',
-        styles: ["\n    /**\n     * Scrollable tbody\n     * https://jsfiddle.net/tsayen/xuvsncr2/28/\n     * http://stackoverflow.com/questions/17067294/html-table-with-100-width-with-vertical-scroll-inside-tbody\n     */\n\n    table {\n      display: flex;\n      flex-flow: column;\n      height: 100%;\n      width: 100%;\n    }\n    table thead {\n      /* head takes the height it requires,\n      and it's not scaled when table is resized */\n      flex: 0 0 auto;\n      width: calc(100% - 0.9em);\n    }\n    table tbody {\n      /* body takes all the remaining available space */\n      flex: 1 1 auto;\n      display: block;\n      overflow: auto;\n    }\n    table tbody tr {\n      width: 100%;\n    }\n    table thead, table tbody tr {\n      display: table;\n      table-layout: fixed;\n    }\n\n    /**\n     * Appearance\n     */\n\n    th span:hover {\n      cursor: pointer;\n    }\n    .table thead th {\n      vertical-align: top;\n      border-bottom: 1px solid #eceeef;\n    }\n    .filter-wrap {\n      margin-top: 8px;\n    }\n\n    /**\n     * Preserve newlines\n     * http://stackoverflow.com/questions/10937218/how-to-show-multiline-text-in-a-table-cell\n     *\n     * And let word-wrap still work\n     * http://stackoverflow.com/a/4413129\n     */\n    td > div > .cell-content {\n      white-space: pre-wrap;\n    }\n  "],
-        template: "\n    <table\n      [ngClass]=\"getNgThing('table', 'class', tableConfig)\"\n      [ngStyle]=\"getNgThing('table', 'style', tableConfig)\">\n      <thead>\n        <tr>\n          <!-- must render from tableConfig so the table don't re-render all the\n               time when tableConfigCopy is updated -->\n          <th *ngFor=\"let col of tableConfigCopy.columnDefs\"\n            [style.width]=\"col.width\"\n            [ngClass]=\"getNgThing('header', 'class', tableConfig, null, null, null, col)\"\n            [ngStyle]=\"getNgThing('header', 'style', tableConfig, null, null, null, col)\">\n            <span (click)=\"colHeaderSortClicked(col)\">\n              {{col.headerText || col.field}}\n            </span>\n            <span *ngIf=\"col.sortAdvanced?.direction === -1\">&#9660;</span>\n            <span *ngIf=\"col.sortAdvanced?.direction === 1\">&#9650;</span>\n\n            <div *ngIf=\"isAnyFieldFilterable\" class=\"filter-wrap\">\n              <filter-input-cmp\n                *ngIf=\"col.filterEnabled\"\n                [filterValue]=\"col.filterValue\"\n                [field]=\"col.field\"\n                [placeholder]=\"col.filterPlaceholder\"\n                (filter)=\"filterUpdate($event)\">\n              </filter-input-cmp>\n            </div>\n          </th>\n        </tr>\n      </thead>\n      <tbody>\n        <tr\n          *ngFor=\"let rowData of tableData | filter: tableConfigCopy; let rowIndex = index\"\n          [ngClass]=\"getNgThing('row', 'class', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow)\"\n          [ngStyle]=\"getNgThing('row', 'style', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow)\"\n          (click)=\"rowClickedFn($event, rowData, rowIndex)\">\n          <td *ngFor=\"let col of tableConfig.columnDefs\" [style.width]=\"col.width\">\n            <div [ngSwitch]=\"col.cellItem?.elementType\">\n\n              <!-- BUTTON -->\n              <div *ngSwitchCase=\"'button'\">\n                <div\n                  [ngClass]=\"getNgThing('cell', 'class', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                  [ngStyle]=\"getNgThing('cell', 'style', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\">\n                  <button type=\"button\"\n                    [ngClass]=\"getNgThing('cellItemButton', 'class', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                    [ngStyle]=\"getNgThing('cellItemButton', 'style', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                    (click)=\"cellItemClickedFn($event, col, rowData)\"\n                    >{{col.cellItem?.staticContent || rowData[col.field]}}</button>\n                </div>\n              </div>\n\n              <!-- DIV -->\n              <div *ngSwitchCase=\"'div'\">\n                <div\n                  [ngClass]=\"getNgThing('cell', 'class', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                  [ngStyle]=\"getNgThing('cell', 'style', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\">\n                  <div\n                    [ngClass]=\"getNgThing('cellItemDiv', 'class', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                    [ngStyle]=\"getNgThing('cellItemDiv', 'style', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                    (click)=\"cellItemClickedFn($event, col, rowData, rowIndex)\"\n                    >{{col.cellItem?.staticContent || rowData[col.field]}}</div>\n                </div>\n              </div>\n\n              <!-- NO ITEM -->\n              <div *ngSwitchDefault class=\"cell-content\"\n                  [ngClass]=\"getNgThing('cell', 'class', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                  [ngStyle]=\"getNgThing('cell', 'style', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                >{{rowData[col.field]}}</div>\n            </div>\n          </td>\n        </tr>\n      </tbody>\n    </table>\n  "
+        styles: ["\n    /**\n     * Scrollable tbody\n     * https://jsfiddle.net/tsayen/xuvsncr2/28/\n     * http://stackoverflow.com/questions/17067294/html-table-with-100-width-with-vertical-scroll-inside-tbody\n     */\n\n    table {\n      display: flex;\n      flex-flow: column;\n      height: 100%;\n      width: 100%;\n    }\n    table thead {\n      /* head takes the height it requires,\n      and it's not scaled when table is resized */\n      flex: 0 0 auto;\n    }\n    table thead.with-scrollbar {\n      width: calc(100% - 0.9em);\n    }\n    table thead.without-scrollbar {\n      width: 100%\n    }\n    table tbody {\n      /* body takes all the remaining available space */\n      flex: 1 1 auto;\n      display: block;\n      overflow: auto;\n    }\n    table tbody tr {\n      width: 100%;\n    }\n    table thead, table tbody tr {\n      display: table;\n      table-layout: fixed;\n    }\n\n    /**\n     * Appearance\n     */\n\n    th span:hover {\n      cursor: pointer;\n    }\n    .table thead th {\n      vertical-align: top;\n      border-bottom: 1px solid #eceeef;\n    }\n    .filter-wrap {\n      margin-top: 8px;\n    }\n\n    /**\n     * Preserve newlines\n     * http://stackoverflow.com/questions/10937218/how-to-show-multiline-text-in-a-table-cell\n     *\n     * And let word-wrap still work\n     * http://stackoverflow.com/a/4413129\n     */\n    td > div > .cell-content {\n      white-space: pre-wrap;\n    }\n  "],
+        template: "\n    <table\n      [ngClass]=\"getNgThing('table', 'class', tableConfig)\"\n      [ngStyle]=\"getNgThing('table', 'style', tableConfig)\">\n      <thead #thead>\n        <tr>\n          <!-- rendering from tableConfigCopy makes the header re-render each\n               time tableConfigCopy is updated -->\n          <th *ngFor=\"let col of tableConfigCopy.columnDefs\"\n            [style.width]=\"col.width\"\n            [ngClass]=\"getNgThing('header', 'class', tableConfig, null, null, null, col)\"\n            [ngStyle]=\"getNgThing('header', 'style', tableConfig, null, null, null, col)\">\n            <span (click)=\"colHeaderSortClicked(col)\">\n              {{col.headerText || col.field}}\n              <span *ngIf=\"col.sortAdvanced?.direction === -1\" class=\"ng2-table-caret\">\n                &nbsp;&#9660;\n              </span>\n              <span *ngIf=\"col.sortAdvanced?.direction === 1\" class=\"ng2-table-caret\">\n                &nbsp;&#9650;\n              </span>\n            </span>\n\n            <div *ngIf=\"!isAnyFieldFilterable\" class=\"filter-wrap\">\n            </div>\n\n            <div *ngIf=\"isAnyFieldFilterable\" class=\"filter-wrap\">\n              <filter-input-cmp\n                *ngIf=\"col.filterEnabled\"\n                [filterValue]=\"col.filterValue\"\n                [field]=\"col.field\"\n                [placeholder]=\"col.filterPlaceholder\"\n                (filter)=\"filterUpdate($event)\">\n              </filter-input-cmp>\n            </div>\n          </th>\n        </tr>\n      </thead>\n      <tbody #tbody>\n        <tr\n          *ngFor=\"let rowData of tableData | filter: tableConfigCopy; let rowIndex = index\"\n          [ngClass]=\"getNgThing('row', 'class', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow)\"\n          [ngStyle]=\"getNgThing('row', 'style', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow)\"\n          (click)=\"rowClickedFn($event, rowData, rowIndex)\">\n          <td *ngFor=\"let col of tableConfig.columnDefs\" [style.width]=\"col.width\">\n            <div [ngSwitch]=\"col.cellItem?.elementType\">\n\n              <!-- BUTTON -->\n              <div *ngSwitchCase=\"'button'\">\n                <div\n                  [ngClass]=\"getNgThing('cell', 'class', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                  [ngStyle]=\"getNgThing('cell', 'style', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\">\n                  <button type=\"button\"\n                    [ngClass]=\"getNgThing('cellItemButton', 'class', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                    [ngStyle]=\"getNgThing('cellItemButton', 'style', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                    (click)=\"cellItemClickedFn($event, col, rowData)\"\n                    >{{col.cellItem?.staticContent || rowData[col.field]}}</button>\n                </div>\n              </div>\n\n              <!-- DIV -->\n              <div *ngSwitchCase=\"'div'\">\n                <div\n                  [ngClass]=\"getNgThing('cell', 'class', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                  [ngStyle]=\"getNgThing('cell', 'style', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\">\n                  <div\n                    [ngClass]=\"getNgThing('cellItemDiv', 'class', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                    [ngStyle]=\"getNgThing('cellItemDiv', 'style', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                    (click)=\"cellItemClickedFn($event, col, rowData, rowIndex)\"\n                    >{{col.cellItem?.staticContent || rowData[col.field]}}</div>\n                </div>\n              </div>\n\n              <!-- NO ITEM -->\n              <div *ngSwitchDefault class=\"cell-content\"\n                  [ngClass]=\"getNgThing('cell', 'class', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                  [ngStyle]=\"getNgThing('cell', 'style', tableConfig, rowData, rowIndex, tableConfigCopy.activeRow, col)\"\n                >{{rowData[col.field]}}</div>\n            </div>\n          </td>\n        </tr>\n      </tbody>\n    </table>\n  "
     })
 ], Ng2TableComponent);
 exports.Ng2TableComponent = Ng2TableComponent;